refactor: move app logic into hook
